Added CQRS

+++ src/Core/ECommerceApp.Application/Dtos/ProductDto.cs
@@ -11,8 +11,8 @@ namespace ECommerceApp.Application.Dtos
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
-        public decimal StockCount { get; set; }
+        public double Price { get; set; }
+        public int StockCount { get; set; }
         public string ImageUrl { get; set; }
         public ProductCategoryDto ProductCategory { get; set; }
     }
